codegen-on-oss/codegen_on_oss/analysis/pr_analysis/requirements.txt
# Core dependencies
PyGithub>=1.55.0
GitPython>=3.1.24
PyYAML>=6.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=3.0.0

